initial commit: Structured Files. - Basic Structure. with pages and helper functions for now.
